Docstrings

Here we give some of the main docstrings.

FVMGeometry

FiniteVolumeMethod.FVMGeometry — Type

FVMGeometry{EL,P,Adj,Adj2V,NGH,TAR,N,BE,IEBEI,T,CT,VS,SS,CS,FT}

Geometric information about the underlying triangular mesh, as needed for the finite volume method.

Fields

mesh_information::MeshInformation

Information about the underlying triangular mesh; see MeshInformation.

boundary_information::BoundaryInformation

Information about the boundary of the underlying triangular mesh; see BoundaryInformation.

interior_information::InteriorInformation

Information about the interior of the underlying triangular mesh; see InteriorInformation.

element_information_list::Dict{T, ElementInformation}

List that maps each element to the information about that mesh; see ElementInformation.

volumes::Dict{Int64,FT}

Maps node indices to the volume of its corresponding control volume.

Constructors

FVMGeometry(tri::Triangulation;
    coordinate_type=Vector{number_type(pts)},
    control_volume_storage_type_vector=NTuple{3,coordinate_type},
    control_volume_storage_type_scalar=NTuple{3,number_type(pts)},
    shape_function_coefficient_storage_type=NTuple{9,number_type(pts)},
    interior_edge_storage_type=NTuple{2,Int64},
    interior_edge_pair_storage_type=NTuple{2,interior_edge_storage_type}) where {Ts}

Constructor for FVMGeometry.

Arguments

tri::Triangulation: The mesh.

Keyword Arguments

coordinate_type=Vector{number_type(pts)}: How coordinates are represented.

control_volume_storage_type_vector=NTuple{3,coordinate_type}: How information for triples of coordinates is represented. The element type must be coordinate_type.

control_volume_storage_type_scalar=NTuple{3,number_type(pts)}: How information triples of scalars is represented. The element type must be the same as the element type used for pts.

shape_function_coefficient_storage_type=NTuple{9,number_type(pts)}: How the nine shape function coefficients for each element are stored. The element type must be the same as the element type used for pts.

interior_edge_storage_type=NTuple{2,Int64}: How the tuples (v, j), as defined for the interior edge identifier in InteriorInformation, are stored.

interior_edge_pair_storage_type=NTuple{2,interior_edge_storage_type}: How the list of tuples for the interior edge tuples, as defined for the interior edge identifier in InteriorInformation, are stored. The element type must be interior_edge_pair_storage_type.

Outputs

The returned value is the FVMGeometry object storing information about the underlying triangular mesh.

FiniteVolumeMethod.MeshInformation — Type

MeshInformation{T<:Triangulation,TAR}

Information about the underlying triangular mesh.

Fields

triangulation::T

The mesh, given with T <: Triangulation.

total_area::TAR

The total area of the mesh.

FiniteVolumeMethod.BoundaryInformation — Type

BoundaryInformation{N,BE}

Information for the boundary of the domain.

Fields

edge_information::BoundaryEdgeMatrix: The BoundaryEdgeMatrix.

normal_information::OutwardNormalBoundary{N}: The OutwardNormalBoundary.

boundary_nodes::Vector{Int64}: The indices for the nodes on the boundary, collected into a single vector.

boundary_elements::BE: The elements of the underlying triangular mesh that share at least one edge with the boundary.

FiniteVolumeMethod.BoundaryEdgeMatrix — Type

BoundaryEdgeMatrix

Information for the boundary edges. Boundaries are to be interpreted according to the specification in DelaunayTriangulation.jl.

Fields

left_nodes::Vector{Int64}: The left nodes for each edge, given as indices for the points.

right_nodes::Vector{Int64}: The right nodes for each edge, given as indices for the points. (Actually, this orientation switches on inner boundaries since those are oriented clockwise rather than counter-clockwise.)

adjacent_nodes::Vector{Int64}: The adjacent node to each edge, i.e. the node such that (left_nodes[i], right_nodes[i], adjacent_nodes[i]) is a positively oriented triangle in the underlying triangular mesh.

types::Vector{Int64}: The boundaries are split into segments, so types[i] says what segment number the ith edge is on.

FiniteVolumeMethod.OutwardNormalBoundary — Type

OutwardNormalBoundary{N}

Information for the unit outward normal vector field on the boundary.

Fields

x_normals::N: The x-coordinates for the tail of each normal vector, with x_normals[i] the coordinate for the ith edge.

y_normals::N: The y-coordinates for the tail of each normal vector, with x_normals[i] the coordinate for the ith edge.

types::Vector{Int64}: As in BoundaryEdgeMatrix, types[i] says what segment number the ith edge is on.

FiniteVolumeMethod.InteriorInformation — Type

InteriorInformation{EL,IEBEI}

Information for the interior of the domain.

Fields

nodes::Vector{Int64}

Indices for the all the nodes that are in the interior of the domain.

elements::EL

The elements of the underlying triangular mesh that share no edges with the boundary.

interior_edge_boundary_element_identifier::IEBEI

Given a boundary element V (see BoundaryInformation), this maps V to a list of tuples of the form ((v₁, j₁), (v₂, j₂)), where (v₁, v₂) is an edge such that v₁ is a boundary node and v₂ is an interior node, and j₁ and j₂ give the positions of v₁ and v₂, respectively, in V, e.g. V[j₁] == v₁. See also construct_interior_edge_boundary_element_identifier.

FiniteVolumeMethod.construct_interior_edge_boundary_element_identifier — Function

construct_interior_edge_boundary_element_identifier(boundary_elements, boundary_edges, tri;
    interior_edge_storage_type, interior_edge_pair_storage_type)

Creates a Dict that maps a set of nodal indices v for some boundary element to the the interior edges, represented as ((v₁, j₁), (v₂, j₂)) (a vector of), with v[j₁] = v₁ → v[j₂] = v₂ an edge of the element that is in the interior.

FiniteVolumeMethod.ElementInformation — Type

ElementInformation{CoordinateType,VectorStorage,ScalarStorage,CoefficientStorage,FloatType}

Information for an individual element.

Fields

centroid::CoordinateType

The centroid of the element.

midpoints::VectorStorage

The coordinates for the midpoints of each edge.

control_volume_edge_midpoints::VectorStorage

For each edge, this is the midpoint of the line that connects the edge to the centroid.

lengths::ScalarStorage

The lengths for each edge.

shape_function_coefficients::CoefficientStorage

These are the sᵢⱼ coefficients for the element; see their definition in the README.

normals::VectorStorage

The outward unit normals to each edge.

area::FloatType

The area of the element.

BoundaryConditions

FiniteVolumeMethod.BoundaryConditions — Type

BoundaryConditions{Tri,F,P,DN,NN,DuN,INN,BMI,MBI,TM}

Information representing the boundary conditions for the PDE.

Fields

triangulation::Tri

The mesh.

functions::F

The Tuple of boundary condition functions for each boundary segment, with functions[i] corresponding to the ith segment. These functions must take the form f(x, y, t, u, p).

parameters::P

The Tuple of arguments p for each boundary condition function, with parameters[i] corresponding to functions[i].

dirichlet_nodes::DN

The indices of the nodes on the boundary that are of Dirichlet type.

neumann_nodes::NN

The indices of the nodes on the boundary that are of Neumann type.

dudt_nodes::DuN

The indices of the nodes on the boundary that are of time-dependent Dirichlet type, i.e. of the form du/dt = f(x, y, t, u, p).

interior_or_neumann_nodes::INN`

The nodes that are either interior or neumann nodes.

boundary_to_mesh_idx::BMI

If the boundary nodes are in the order (b₁, b₂, …, bᵢ, …), then this is a map that takes the index order i to the corresponding index in the mesh, i.e. to bᵢ.

mesh_to_boundary_idx::MBI

The inverse map of boundary_to_mesh_idx.

type_map::TM

Given a node, maps it to the segment number that it belongs to.

Constructors

BoundaryConditions(mesh::FVMGeometry, functions, types;
    params=Tuple(nothing for _ in (functions isa Function ? [1] : eachindex(functions))),
    u_type=Float64, float_type=Float64)

Constructor for the BoundaryConditions struct.

Arguments

mesh::FVMGeometry: The FVMGeometry for the mesh.

functions: The functions for each boundary segment, taking the forms f(x, y, t, u, p). Can be a single function, doesn't have to be in a container (as long as only one segment is needed).

types: The classification for the boundary condition type on each segment. See is_dirichlet_type, is_neumann_type, and is_dudt_type for the possible values here. types[i] is the classification for the ith segment.

Keyword Arguments

params=Tuple(nothing for _ in (functions isa Function ? [1] : eachindex(functions))): The parameters for the functions, with params[i] giving the argument p in functions[i].

u_type=Float64: The number type used for the solution.

float_type=Float64: The number type used for representing the coordinates of points.

Outputs

The returned value is the corresponding BoundaryConditions struct.

FiniteVolumeMethod.is_dirichlet_type — Function

is_dirichlet_type(type)

Returns type ∈ (:Dirichlet, :D, :dirichlet, "Dirichlet", "D", "dirichlet").

FiniteVolumeMethod.is_neumann_type — Function

is_neumann_type(type)

Returns type ∈ (:Neumann, :N, :neumann, "Neumann", "N", "neumann").

FiniteVolumeMethod.is_dudt_type — Function

is_dudt_type(type)

Returns type ∈ (:Dudt, :dudt, "Dudt", "dudt", "du/dt").

FVMProblem

FiniteVolumeMethod.FVMProblem — Type

FVMProblem{iip_flux,FG,BC,F,FP,R,RP,IC,FT}

Struct representing the PDE problem.

Fields

mesh::FG

The underlying mesh, given as a FVMGeometry struct.

boundary_conditions::BC

The boundary conditions, given a a BoundaryConditions struct.

flux_function::F

The function for the flux vector, taking either of the forms flux!(q, x, y, t, α, β, γ, p) (if iip_flux) and flux(x, y, t, α, β, γ, p) (if !iip_flux), where u = αx + βy + γ.

flux_parameters::FP

The argument p in flux_function.

reaction_function::R

The function for the reaction term, taking the form R(x, y, t, u, p).

reaction_parameters::RP

The argument p in reaction_function.

initial_condition::IC

The initial condition for the problem, with initial_condition[i] the initial value at the ith node of the mesh.

initial_time::FT

The time to start solving the PDE at.

final_time::FT

The time to stop solving the PDE at.

steady::Bool

Whether ∂u/∂t = 0 or not. When this is true, the problem is defined as a NonlinearProblem with initial guess coming from the initial condition above.

Constructor

FVMProblem(mesh, boundary_conditions;
    iip_flux=true,
    diffusion_function=nothing,
    diffusion_parameters=nothing,
    reaction_function=nothing,
    reaction_parameters=nothing,
    delay_function=nothing,
    delay_parameters=nothing,
    flux_function=nothing,
    flux_parameters=nothing,
    initial_condition,
    initial_time=0.0,
    final_time,
    steady=false,
    q_storage=SVector{2,Float64})

Constructor for the FVMProblem.

Arguments

mesh: The FVMGeometry representing the underlying mesh.

boundary_conditions: The BoundaryConditions representing the boundary conditions for the problem.

Keyword Arguments

iip_flux=true: Whether the flux vector is computed in-place or not.

diffusion_function=nothing: If flux_function===nothing, this can be used to provide a diffusion term for the reaction-diffusion formulation, taking the form D(x, y, t, u, p). See also construct_flux_function.

diffusion_parameters=nothing: The argument p in the diffusion function.

reaction_function=nothing: The reaction term, taking the form R(x, y, t, u, p). If not provided, it is set to the zero function. See also construct_reaction_function.

reaction_parameters=nothing: The argument p in the reaction function.

delay_function=nothing: The delay function T(x, y, t, u, p) for the PDE that can be used to scale the diffusion and reaction functions, assuming flux_function===nothing. See also construct_reaction_function and construct_flux_function.

delay_parameters=nothing: The argument p in the delay function.

flux_function=nothing: The flux function, taking either of the forms flux!(q, x, y, t, α, β, γ, p) (if iip_flux) and flux(x, y, t, α, β, γ, p) (if !iip_flux), where u = αx + βy + γ. If flux_function===nothing, thne this function is constructed from the diffusion and delay functions.

flux_parameters=nothing: The argument p in the flux function.

initial_condition: The initial condition for the problem, with initial_condition[i] the initial value at the ith node of the mesh.

initial_time=0.0: The time to start solving the PDE at.

final_time: The time to stop solving the PDE at.

steady::Bool: Whether ∂u/∂t = 0 or not. If the problem is steady, then the initial estimate used for the nonlinear solver that finds the steady state is given by the initial condition, and final_time is set to ∞.

Outputs

Returns the FVMProblem object.

FiniteVolumeMethod.construct_flux_function — Function

construct_flux_function(iip_flux,
    flux_function,
    delay_function, delay_parameters,
    diffusion_function, diffusion_parameters)

Constructs the flux function. The arguments are as in FVMProblem, and the output depends on the following, where D denotes the diffusion function, T the delay function, dₚ the diffusion parameters, and tₚ the delay parameters:

If flux_function===nothing and delay_function===nothing, defines the flux function as (x, y, t, α, β, γ, _) -> -D(x, y, t, αx + βy + γ, dₚ)[α, β].

If flux_function===nothing, defines the flux function as (x, y, t, α, β, γ, _) -> -T(x, y, t, αx + βy + γ, tₚ)D(x, y, t, αx + βy + γ, dₚ)[α, β].

If iip_flux, then the functions above have an extra argument in the first position, q, that the flux vectors are stored in. Otherwise, the flux vector is returned as a tuple (q1, q2). If flux_function !== nothing, then just returns flux_function.

FiniteVolumeMethod.construct_reaction_function — Function

construct_reaction_function(reaction_function, reaction_parameters,
    delay_function, delay_parameters)

Constructs the reaction function. The arguments are as in FVMProblem, and the output depends on the following, where R denotes the reaction function, T the delay function, rₚ the reaction parameters, and tₚ the delay parameters:

If reaction_function===nothing, defines the reaction function as (x, y, t, u, _) = 0.0.

If reaction_function !== nothing and delay_function !== nothing, defines the reaction function as (x, y, t, u, _) -> T(x, y, t, u, tₚ)R(x, y, t, u, rₚ).

Otherwise, just returns reaction_function.

Solving the FVMProblem

CommonSolve.solve — Function

SciMLBase.solve(prob::FVMProblem, alg;
    cache_eltype::Type{F}=eltype(get_initial_condition(prob)),
    jac_prototype=float.(jacobian_sparsity(prob)),
    parallel=false,
    specialization::Type{S}=SciMLBase.AutoSpecialize,
    chunk_size=PreallocationTools.ForwardDiff.pickchunksize(length(get_initial_condition(prob))),
    kwargs...)

Solves the FVMProblem given by prob using the algorithm alg.

Arguments

prob::FVMProblem: The FVMProblem.

alg: The algorithm to use for solving. See the DifferentialEquations.jl documentation for this. If your problem is a steady problem, then you should use an algorithm from NonlinearSolve.jl instead - note that the initial estimate in this case comes from the initial condition.

Keyword Arguments

cache_eltype::Type{F}=eltype(get_initial_condition(prob)): The element type used for the cache vectors.

jac_prototype=float.(jacobian_sparsity(prob)): The prototype for the sparsity pattern of the Jacobian.

parallel=false: Whether to use multithreading for evaluating the equations. Not currently used.

specialization::Type{S}=SciMLBase.AutoSpecialize: The specialisation level for the ODEProblem.

chunk_size=PreallocationTools.ForwardDiff.pickchunksize(length(get_initial_condition(prob)))): The chunk size for the dual numbers used in the cache vector.

kwargs...: Extra keyword arguments for solve as used on ODEProblems (or NonlinearProblems if your problem is steady).

Output

The output is a solution struct, as returned from DifferentialEquations.jl. If instead your problem is steady, returns a solution struct as returned from NonlinearSolve.jl.

SciMLBase.ODEProblem — Type

SciMLBase.ODEProblem(prob::FVMProblem;
    cache_eltype::Type{F}=eltype(get_initial_condition(prob)),
    jac_prototype=float.(jacobian_sparsity(prob)),
    parallel=false,
    no_saveat=true,
    specialization::Type{S}=SciMLBase.AutoSpecialize,
    chunk_size=PreallocationTools.ForwardDiff.pickchunksize(length(get_initial_condition(prob)))) where {S,F}

Constructs the ODEProblem for the system of ODEs defined by the finite volume equations corresponding to prob.

Arguments

prob::FVMProblem: The FVMProblem.

Keyword Arguments

cache_eltype::Type{F}=eltype(get_initial_condition(prob)): The element type used for the cache vectors.

jac_prototype=float.(jacobian_sparsity(prob)): The prototype for the sparsity pattern of the Jacobian.

parallel=false: Whether to use multithreading for evaluating the equations.

no_saveat=true: Whether the solution is saving at specific points.

specialization::Type{S}=SciMLBase.AutoSpecialize: The specialisation level for the ODEProblem.

chunk_size=PreallocationTools.ForwardDiff.pickchunksize(length(get_initial_condition(prob)))): The chunk size for the dual numbers used in the cache vector.

Outputs

Returns the corresponding ODEProblem.

FiniteVolumeMethod.jacobian_sparsity — Function

jacobian_sparsity(prob::FVMProblem)

Constructs the sparse matrix which has the same sparsity pattern as the Jacobian for the finite volume equations corresponding to the FVMProblem given by prob.

Linear Interpolants

FiniteVolumeMethod.eval_interpolant — Function

eval_interpolant!(αβγ, prob::FVMProblem, x, y, T, u)
eval_interpolant(sol, x, y, t_idx::Integer, T)
eval_interpolant(sol, x, y, t::Number, T)

Evaluates the interpolant corresponding to the FVMProblem at the point (x, y) and time t (or sol.t[t_idx]).